+++ R/rsi_calc.R
@@ -148,7 +148,9 @@ rsi_calc <- function(...,
   
   if (print_warning == TRUE) {
     if (message_not_thrown_before("rsi_calc")) {
-      warning_("Increase speed by transforming to class <rsi> on beforehand: your_data %>% mutate_if(is.rsi.eligible, as.rsi)",
+      warning_("Increase speed by transforming to class <rsi> on beforehand:\n",
+               "  your_data %>% mutate_if(is.rsi.eligible, as.rsi)\n",
+               "  your_data %>% mutate(across((is.rsi.eligible), as.rsi))",
                call = FALSE)
       remember_thrown_message("rsi_calc")
     }